backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
sqlalchemy
pydantic
pydantic-settings
